Drop requirement for Rust nightly
