+++ Cargo.toml
@@ -35,3 +35,13 @@ owning_ref = "0.4.1"
 tempfile = "3.23.0"
 test_bin = "0.5.0"
 unicode-segmentation = "1.12.0"
+
+[dev-dependencies]
+criterion = "0.7.0"
+
+[lib]
+bench = false
+
+[[bench]]
+name = "roundtrip"
+harness = false
